requirements.txt
streamlit>=1.10.0
mysql-connector-python>=8.0.26
streamlit>=1.30.0
pandas>=1.5.0
scikit-learn>=1.2.0
joblib>=1.2.0
numpy>=1.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
